Moved DrlxParseResult to separated package
Renamed drlxParse to ConstraintParser

Constraint Parser is a class with context, packageModel

Introducing two cases for failure
