+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/WindowReferenceGenerator.java
@@ -24,10 +24,14 @@ import org.drools.javaparser.ast.type.Type;
 import org.drools.model.Window;
 import org.drools.model.WindowDefinition;
 import org.drools.modelcompiler.builder.PackageModel;
+import org.drools.modelcompiler.builder.generator.drlxparse.DrlxParseFail;
+import org.drools.modelcompiler.builder.generator.drlxparse.DrlxParseResult;
+import org.drools.modelcompiler.builder.generator.drlxparse.DrlxParseSuccess;
+import org.drools.modelcompiler.builder.generator.drlxparse.ConstraintParser;
+import org.drools.modelcompiler.builder.generator.drlxparse.ParseResultVisitor;
 
 import static org.drools.modelcompiler.builder.generator.DrlxParseUtil.generateLambdaWithoutParameters;
 import static org.drools.modelcompiler.builder.generator.DrlxParseUtil.toVar;
-import static org.drools.modelcompiler.builder.generator.ModelGenerator.drlxParse;
 
 public class WindowReferenceGenerator {
 
@@ -90,12 +94,22 @@ public class WindowReferenceGenerator {
     }
 
     private Optional<Expression> parseCondition( KnowledgeBuilderImpl kbuilder, PackageModel packageModel, PatternDescr pattern, Class<?> patternType ) {
-        return Optional.ofNullable(pattern.getConstraint().getDescrs().iterator().next()).map(d -> {
+        return Optional.ofNullable(pattern.getConstraint().getDescrs().iterator().next()).flatMap(d -> {
             String expression = d.toString();
             RuleContext context = new RuleContext(kbuilder, pkg, packageModel.getExprIdGenerator(), null);
-            DrlxParseResult drlxParseResult = drlxParse(context, packageModel, patternType, pattern.getIdentifier(), expression);
+            final DrlxParseResult drlxParseResult = new ConstraintParser(context, packageModel).drlxParse(patternType, pattern.getIdentifier(), expression);
 
-            return generateLambdaWithoutParameters(drlxParseResult.getUsedDeclarations(), drlxParseResult.getExpr());
+            return drlxParseResult.acceptWithReturnValue(new ParseResultVisitor<Optional<Expression>>() {
+                @Override
+                public Optional<Expression> onSuccess(DrlxParseSuccess drlxParseResult) {
+                    return Optional.of(generateLambdaWithoutParameters(drlxParseResult.getUsedDeclarations(), drlxParseResult.getExpr()));
+                }
+
+                @Override
+                public Optional<Expression> onFail(DrlxParseFail failure) {
+                    return Optional.empty();
+                }
+            });
         });
     }
