+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/drlxparse/DrlxParseFail.java
@@ -1,0 +1,14 @@
+package org.drools.modelcompiler.builder.generator.drlxparse;
+
+public class DrlxParseFail implements DrlxParseResult {
+
+    @Override
+    public void accept(ParseResultVoidVisitor parseVisitor) {
+        parseVisitor.onFail(this);
+    }
+
+    @Override
+    public <T> T acceptWithReturnValue(ParseResultVisitor<T> visitor) {
+        return visitor.onFail(this);
+    }
+}
