+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/drlxparse/DrlxParseResult.java
@@ -1,0 +1,9 @@
+package org.drools.modelcompiler.builder.generator.drlxparse;
+
+public interface DrlxParseResult {
+
+    void accept(ParseResultVoidVisitor visitor);
+
+    <T> T acceptWithReturnValue(ParseResultVisitor<T> visitor);
+
+}
