+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/drlxparse/ParseResultVisitor.java
@@ -1,0 +1,8 @@
+package org.drools.modelcompiler.builder.generator.drlxparse;
+
+public interface ParseResultVisitor<T> {
+    T onSuccess(DrlxParseSuccess drlxParseResult);
+
+    T onFail(DrlxParseFail failure);
+
+}
