+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/drlxparse/ParseResultVoidVisitor.java
@@ -1,0 +1,11 @@
+package org.drools.modelcompiler.builder.generator.drlxparse;
+
+@FunctionalInterface
+public interface ParseResultVoidVisitor {
+    void onSuccess(DrlxParseSuccess drlxParseResult);
+
+    default void onFail(DrlxParseFail failure) {
+
+    }
+
+}
